dataset/python-mutated/test_geography.py
from superset.utils.pandas_postprocessing import geodetic_parse, geohash_decode, geohash_encode
from tests.unit_tests.fixtures.dataframes import lonlat_df
from tests.unit_tests.pandas_postprocessing.utils import round_floats, series_to_list

def test_geohash_decode():
    if False:
        while True:
            i = 10
    post_df = geohash_decode(df=lonlat_df[['city', 'geohash']], geohash='geohash', latitude='latitude', longitude='longitude')
    assert sorted(post_df.columns.tolist()) == sorted(['city', 'geohash', 'latitude', 'longitude'])
    assert round_floats(series_to_list(post_df['longitude']), 6) == round_floats(series_to_list(lonlat_df['longitude']), 6)
    assert round_floats(series_to_list(post_df['latitude']), 6) == round_floats(series_to_list(lonlat_df['latitude']), 6)

def test_geohash_encode():
    if False:
        while True:
            i = 10
    post_df = geohash_encode(df=lonlat_df[['city', 'latitude', 'longitude']], latitude='latitude', longitude='longitude', geohash='geohash')
    assert sorted(post_df.columns.tolist()) == sorted(['city', 'geohash', 'latitude', 'longitude'])
    assert series_to_list(post_df['geohash']) == series_to_list(lonlat_df['geohash'])

def test_geodetic_parse():
    if False:
        return 10
    post_df = geodetic_parse(df=lonlat_df[['city', 'geodetic']], geodetic='geodetic', latitude='latitude', longitude='longitude', altitude='altitude')
    assert sorted(post_df.columns.tolist()) == sorted(['city', 'geodetic', 'latitude', 'longitude', 'altitude'])
    assert series_to_list(post_df['longitude']) == series_to_list(lonlat_df['longitude'])
    assert series_to_list(post_df['latitude']) == series_to_list(lonlat_df['latitude'])
    assert series_to_list(post_df['altitude']) == series_to_list(lonlat_df['altitude'])
    post_df = geodetic_parse(df=lonlat_df[['city', 'geodetic']], geodetic='geodetic', latitude='latitude', longitude='longitude')
    assert sorted(post_df.columns.tolist()) == sorted(['city', 'geodetic', 'latitude', 'longitude'])
    assert series_to_list(post_df['longitude']) == series_to_list(lonlat_df['longitude'])
    assert series_to_list(post_df['latitude']), series_to_list(lonlat_df['latitude'])